+++ bevy_renet2/src/transport.rs
@@ -7,7 +7,7 @@ use bevy_app::{prelude::*, AppExit};
 use bevy_ecs::prelude::*;
 use bevy_time::prelude::*;
 
-use crate::{client_disconnected, RenetClientPlugin, RenetReceive, RenetSend, RenetServerPlugin};
+use crate::{client_should_update, RenetClientPlugin, RenetReceive, RenetSend, RenetServerPlugin};
 
 pub struct NetcodeServerPlugin;
 
@@ -71,14 +71,14 @@ impl Plugin for NetcodeClientPlugin {
             Self::update_system
                 .in_set(RenetReceive)
                 .run_if(resource_exists::<NetcodeClientTransport>)
-                .run_if(not(client_disconnected))
+                .run_if(client_should_update())
                 .after(RenetClientPlugin::update_system),
         );
         app.add_systems(
             PostUpdate,
             (Self::send_packets.in_set(RenetSend), Self::disconnect_on_exit)
                 .run_if(resource_exists::<NetcodeClientTransport>)
-                .run_if(not(client_disconnected)),
+                .run_if(client_should_update()),
         );
     }
 }